+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.\
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines
 and [Semantic Versioning](https://semver.org/spec/v2.0.0.html) for versioning adherance.
 
-## [💾](https://github.com/kyaulabs/grimoire/releases/tag/v0.1.0) [0.1.0](https://github.com/kyaulabs/grimoire/commit/765a95bdefd4fc63f44fd33001669e11365947fc) - (2023-05-22)
+### ♻️ Refactor
+
+- [Tests] Test naming scheme and proper format ([7ef7104](https://github.com/kyaulabs/grimoire/commit/7ef7104f606a5bbdbe82224af31fed2467591058))
+<!-- Auto-generated on 2023-05-22T16:44:23.457594100-07:00 -->
+
+## [💾](https://github.com/kyaulabs/grimoire/releases/tag/v0.1.0) [0.1.0](https://github.com/kyaulabs/grimoire/commit/05b03fc749e4bc3b4f3a19974564be5e3e69a021) - (2023-05-22)
 
 ### ⚗️ Testing
 - Initial test environment setup ([c6583b6](https://github.com/kyaulabs/grimoire/commit/c6583b6724169ba6215f56dac2c75eee0daa2217))
@@ -12,5 +17,5 @@ and [Semantic Versioning](https://semver.org/spec/v2.0.0.html) for versioning ad
 
 ### 🚀 Features
 
-- [Editor] Automated project setup ([765a95b](https://github.com/kyaulabs/grimoire/commit/765a95bdefd4fc63f44fd33001669e11365947fc))
-<!-- Auto-generated on 2023-05-22T13:12:06.797801700-07:00 -->
+- [Editor] Automated project setup ([05b03fc](https://github.com/kyaulabs/grimoire/commit/05b03fc749e4bc3b4f3a19974564be5e3e69a021))
+<!-- Auto-generated on 2023-05-22T16:44:23.458006700-07:00 -->